- second critical fix to broken sitter rank logic.  (harmful newline was throwing off the calculation) - added unit test for sitter rank scoring, from 0 to 12 sittings - fixed data typing error on user.review_ratings_sum - added Excel spreadsheet I used to cobble together the seed.rb creation calls

+++ db/migrate/20140422062725_create_initial_schemas.rb
@@ -51,7 +51,7 @@ class CreateInitialSchemas < ActiveRecord::Migration
             t.float :ratings_score, default: 0.0
             t.float :sitter_rank, index: true, default: 0.0
 
-            t.float :review_ratings_sum, default: 0.0
+            t.integer :review_ratings_sum, default: 0
             t.integer :num_sitter_stays, default: 0
 
             t.timestamps
